MachineLearning/LLM/Project/requirements.txt
Flask
numpy
python-dotenv
llama-index